Add INTEGER support to Hive connector
Change Presto type corresponding to HIVE_INT, HIVE_SMALLINT, and
HIVE_TINYINT to be INTEGER.

+++ presto-hive/src/main/java/com/facebook/presto/hive/HiveBucketing.java
@@ -51,6 +51,7 @@ import static com.facebook.presto.spi.type.BigintType.BIGINT;
 import static com.facebook.presto.spi.type.BooleanType.BOOLEAN;
 import static com.facebook.presto.spi.type.DateType.DATE;
 import static com.facebook.presto.spi.type.DoubleType.DOUBLE;
+import static com.facebook.presto.spi.type.IntegerType.INTEGER;
 import static com.facebook.presto.spi.type.TimestampType.TIMESTAMP;
 import static com.facebook.presto.spi.type.VarcharType.VARCHAR;
 import static com.google.common.base.MoreObjects.toStringHelper;
@@ -118,11 +119,11 @@ final class HiveBucketing
                     case BOOLEAN:
                         return BOOLEAN.getBoolean(block, position) ? 1 : 0;
                     case BYTE:
-                        return SignedBytes.checkedCast(BIGINT.getLong(block, position));
+                        return SignedBytes.checkedCast(INTEGER.getLong(block, position));
                     case SHORT:
-                        return Shorts.checkedCast(BIGINT.getLong(block, position));
+                        return Shorts.checkedCast(INTEGER.getLong(block, position));
                     case INT:
-                        return Ints.checkedCast(BIGINT.getLong(block, position));
+                        return Ints.checkedCast(INTEGER.getLong(block, position));
                     case LONG:
                         long bigintValue = BIGINT.getLong(block, position);
                         return (int) ((bigintValue >>> 32) ^ bigintValue);
